refactor(auth): Move relevant functions to ProposalAuth
