+++ src/mutations/ProposalMutations.ts
@@ -33,6 +33,7 @@ import { AdministrationProposalArgs } from '../resolvers/mutations/Administratio
 import { ChangeProposalsStatusInput } from '../resolvers/mutations/ChangeProposalsStatusMutation';
 import { CloneProposalsInput } from '../resolvers/mutations/CloneProposalMutation';
 import { UpdateProposalArgs } from '../resolvers/mutations/UpdateProposalMutation';
+import { ProposalAuthorization } from './../auth/ProposalAuthorization';
 import { CallDataSource } from './../datasources/CallDataSource';
 import { ProposalSettingsDataSource } from './../datasources/ProposalSettingsDataSource';
 import { CloneUtils } from './../utils/CloneUtils';
@@ -40,6 +41,7 @@ import { CloneUtils } from './../utils/CloneUtils';
 @injectable()
 export default class ProposalMutations {
   private userAuth = container.resolve(UserAuthorization);
+  private proposalAuth = container.resolve(ProposalAuthorization);
   private cloneUtils = container.resolve(CloneUtils);
   constructor(
     @inject(Tokens.ProposalDataSource)
@@ -119,7 +121,7 @@ export default class ProposalMutations {
 
     if (
       !this.userAuth.isUserOfficer(agent) &&
-      !(await this.userAuth.isMemberOfProposal(agent, proposal))
+      !(await this.proposalAuth.isMemberOfProposal(agent, proposal))
     ) {
       return rejection('Unauthorized proposal update', { args });
     }
@@ -204,7 +206,7 @@ export default class ProposalMutations {
     const isUserOfficer = this.userAuth.isUserOfficer(agent);
     if (
       !isUserOfficer &&
-      !(await this.userAuth.isMemberOfProposal(agent, proposal))
+      !(await this.proposalAuth.isMemberOfProposal(agent, proposal))
     ) {
       return rejection('Unauthorized submission of the proposal', {
         agent,
@@ -267,7 +269,7 @@ export default class ProposalMutations {
     if (!this.userAuth.isUserOfficer(agent)) {
       if (
         proposal.submitted ||
-        !this.userAuth.isPrincipalInvestigatorOfProposal(agent, proposal)
+        !this.proposalAuth.isPrincipalInvestigatorOfProposal(agent, proposal)
       )
         return rejection(
           'Can not delete proposal because proposal is submitted',
@@ -335,7 +337,7 @@ export default class ProposalMutations {
       managementDecisionSubmitted,
     } = args;
     const isChairOrSecretaryOfProposal =
-      await this.userAuth.isChairOrSecretaryOfProposal(agent, primaryKey);
+      await this.proposalAuth.isChairOrSecretaryOfProposal(agent, primaryKey);
     const isUserOfficer = this.userAuth.isUserOfficer(agent);
 
     if (!isChairOrSecretaryOfProposal && !isUserOfficer) {
@@ -469,7 +471,7 @@ export default class ProposalMutations {
       );
     }
 
-    if (!(await this.userAuth.hasAccessRights(agent, sourceProposal))) {
+    if (!(await this.proposalAuth.hasAccessRights(agent, sourceProposal))) {
       return rejection(
         'Can not clone proposal because of insufficient permissions',
         { sourceProposal, agent }
