+++ src/mutations/SEPMutations.ts
@@ -10,6 +10,7 @@ import {
 } from '@esss-swap/duo-validation';
 import { container, inject, injectable } from 'tsyringe';
 
+import { ProposalAuthorization } from '../auth/ProposalAuthorization';
 import { UserAuthorization } from '../auth/UserAuthorization';
 import { Tokens } from '../config/Tokens';
 import { InstrumentDataSource } from '../datasources/InstrumentDataSource';
@@ -44,6 +45,7 @@ import { UpdateSEPTimeAllocationArgs } from '../resolvers/mutations/UpdateSEPPro
 @injectable()
 export default class SEPMutations {
   private userAuth = container.resolve(UserAuthorization);
+  private proposalAuth = container.resolve(ProposalAuthorization);
   constructor(
     @inject(Tokens.SEPDataSource)
     private dataSource: SEPDataSource,
@@ -395,7 +397,10 @@ export default class SEPMutations {
     args: SaveSEPMeetingDecisionInput
   ): Promise<SepMeetingDecision | Rejection> {
     const isChairOrSecretaryOfProposal =
-      await this.userAuth.isChairOrSecretaryOfProposal(agent, args.proposalPk);
+      await this.proposalAuth.isChairOrSecretaryOfProposal(
+        agent,
+        args.proposalPk
+      );
     const isUserOfficer = this.userAuth.isUserOfficer(agent);
 
     if (!isChairOrSecretaryOfProposal && !isUserOfficer) {
